# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação Microserviço agenda médica",
        "description": "Swagger Microserviço agenda médica",
        "contact": {
            "email": "1094820@sga.pucminas.br"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/agenda": {
            "get": {
                "tags": [
                    "Agenda"
                ],
                "summary": "Retorna todas as agendas",
                "description": "Retorna todas as agendas",
                "operationId": "97a759a3056cfc72c31a7613d6a9d082",
                "responses": {
                    "200": {
                        "description": "Cadastrado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Dados de retorno",
                                        "value": {
                                            "id": "1",
                                            "data": "13/10/2023"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Não autenticado",
                                        "value": {
                                            "message": "unauthenticated"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Mensagem de erro da exception",
                                        "value": {
                                            "message": "Exemplo de mensagem de erro"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Agenda"
                ],
                "summary": "Cadastra nova agenda",
                "description": "Cadastra nova agenda",
                "operationId": "971a3fd96118d11e0939c5881f73a3ae",
                "requestBody": {
                    "description": "Dados da agenda",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AgendaMedica"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cadastrado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Salvo com sucesso",
                                        "value": {
                                            "message": "Exemplo de mensagem de sucesso"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Não autenticado",
                                        "value": {
                                            "message": "unauthenticated"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Entrada de dados inválida.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Dados inválidos",
                                        "value": {
                                            "message": "Dados inválidos."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Mensagem de erro da exception",
                                        "value": {
                                            "message": "Exemplo de mensagem de erro"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_token": []
                    }
                ]
            }
        },
        "/agenda/{id}": {
            "get": {
                "tags": [
                    "Agenda"
                ],
                "summary": "Retorna agenda pelo id",
                "description": "Retorna agenda pelo id",
                "operationId": "2e16cafcd019777aa8b41e780b53c710",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id da agenda",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Salvo com sucesso",
                                        "value": {
                                            "message": "Exemplo de mensagem de sucesso"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Não autenticado",
                                        "value": {
                                            "message": "unauthenticated"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Entrada de dados inválida.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Dados inválidos",
                                        "value": {
                                            "message": "Dados inválidos."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Mensagem de erro da exception",
                                        "value": {
                                            "message": "Exemplo de mensagem de erro"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Agenda"
                ],
                "summary": "Edita agenda",
                "description": "Edita agenda",
                "operationId": "0d04f2c5d2d08891bacd518c0e187ea7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id da agenda",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "description": "Dados da agenda a serem atualizados",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AgendaMedica"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Salvo com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Salvo com sucesso",
                                        "value": {
                                            "message": "Exemplo de mensagem de sucesso"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Não autenticado",
                                        "value": {
                                            "message": "unauthenticated"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Entrada de dados inválida.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Dados inválidos",
                                        "value": {
                                            "message": "Dados inválidos."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Mensagem de erro da exception",
                                        "value": {
                                            "message": "Exemplo de mensagem de erro"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Agenda"
                ],
                "summary": "Apaga registro da agenda",
                "description": "Apaga registro da agenda",
                "operationId": "bb8d1fd952a7d2168e34a959a203358a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id da agenda",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Apagado com sucesso."
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Não autenticado",
                                        "value": {
                                            "message": "unauthenticated"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Entrada de dados inválida.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Dados inválidos",
                                        "value": {
                                            "message": "Dados inválidos."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Mensagem de erro da exception",
                                        "value": {
                                            "message": "Exemplo de mensagem de erro"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AgendaMedica": {
                "title": "Agenda Médica",
                "properties": {
                    "data": {
                        "type": "string",
                        "format": "date"
                    },
                    "hora": {
                        "type": "string"
                    },
                    "observacoes": {
                        "type": "string"
                    },
                    "cpf_paciente": {
                        "type": "string"
                    },
                    "tipo_consulta": {
                        "type": "string"
                    },
                    "fk_medico_id": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "api_token": {
                "type": "apiKey",
                "description": "Entre com o token beraer no formato **Bearer &lt;token>**",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}